Move getDisplayValue to helpers
